Dialogue Canvas Adjustments
(We are so back)
Dialogue_Canvas prefab now looks lie the demo one (minor adjustments to positioning were made all around).
Header panel gets removed when no one is speaking.
For john's sake, note that Sariel's light has a sprite renderer so the code doesn't yell at me (but I deactivated it)
ALSO, renamed some of the script subfiles to not be redundant.